Refactor `Platforms/WASI/__main__.py` for future subcommands Along the way, support lazy importing.
